+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability for user to provide a preference or `None` to the `ERVD` and `WASPAS` ranking methods. The the user provides `None`, the ranking methods default to the recommended default values.
 - CILOS weighting method. Adding this also required adding an additional error variant, `ValidationError::SingularMatrix`, in the event a certain matrix isn't a square matrix when it actually should be.
 - IDOCRIW weighting method.
+- D-CRITIC weighting method.
+- New `correlation` module for computing correlation coefficients of matrices.
 
 ### Changed